+++ CHANGELOG.md
@@ -2,11 +2,11 @@
 
 ## Pending
 
----
-
 ## 1.1.1
 
 - Added support for `Carthage`, which allows project to use TABTestKit using `Carthage` as their dependency manager.
+
+---
 
 ## 1.1.0