+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 - Clears launch arguments and launch environment values as part of tearDown, ready for the next test.
 - Added new `ActivitySheet` to represent the `UIActivityViewController` sheet.
 - Added support for `Carthage`, which allows project to use TABTestKit using `Carthage` as their dependency manager.
+- Renames `func keyboard(isType type: Keyboard.KeyboardType)` to `func keyboardType(is type: Keyboard.KeyboardType)` to avoid conflict with `public let keyboard = Keyboard()` within `KeyboardContext`.
 
 ---